Revamp the article about notation.
Several changes in the article: there is now a proper set for bit vectors,
and also one for Booleans. The repr function is formally defined to
associate a bit vector to any element of B or G. The article is divided in
subsections. An unary bitwise NOT operator is added to the article. The
sets G+ and G- are infinite and not bounded by an undefined n anymore.